backend/requirements.txt
fastapi
uvicorn[standard]
requests
pandas
numpy
scikit-learn
python-dotenv
google-generativeai
fastapi[all]
